Rename portlist to interface browser
