+++ etc/toolbox/interface_browser.tool
@@ -1,0 +1,4 @@
+name=Interface browser
+uri=/interface_browser/
+icon=/static/images/toolbox/info.png
+description=View detailed interface information for multiple interfaces
