--- etc/toolbox/portlist.tool
@@ -1,4 +1,0 @@
-name=PortList
-uri=/portlist/
-icon=/static/images/toolbox/info.png
-description=View detailed port information for multiple ports
