+++ htdocs/static/js/src/interface_browser_column_toggler.js
