+++ htdocs/static/js/src/interface_browser_dynamic_columns_controller.js
