+++ htdocs/static/js/src/interface_browser_filter_controller.js
