+++ python/nav/web/interface_browser/urls.py
@@ -12,27 +12,13 @@
 # more details.  You should have received a copy of the GNU General Public
 # License along with NAV. If not, see <http://www.gnu.org/licenses/>.
 #
-"""Controllers and views for PortList tool"""
+"""URL-config for PortList tool"""
 
-from django.shortcuts import render
-
-from nav.web import utils
-
-
-def default_context():
-    navpath = (('Home', '/'), ('PortList',))
-    return {
-        'navpath': navpath,
-        'title': utils.create_title(navpath)
-    }
+from django.conf.urls import url
+from nav.web.interface_browser import views
 
 
-def index(request):
-    context = default_context()
-    return render(request, 'portlist/base.html', context)
-
-
-def by_netboxid(request, netboxid):
-    context = default_context()
-    context.update({'netboxid': netboxid})
-    return render(request, 'portlist/base.html', context)
+urlpatterns = [
+    url(r'^$', views.index),
+    url(r'^(?P<netboxid>\d+)', views.by_netboxid, name='interface_browser-netbox')
+]
